pyproject.toml
[project]
name = "langgraph-mem0-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "mem0ai>=0.1.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.5.3",
    "chromadb>=1.0.16",
]
